TodoList 수정
List -> TodoList 이름 수정
TodoItem 컴포넌트로 추출
guard clauses 적용
